Update Avatar class.
